+++ src/main/java/com/onlyoffice/manager/url/DefaultUrlManager.java
@@ -27,7 +27,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.http.client.utils.URIBuilder;
+import org.apache.hc.core5.net.URIBuilder;
 
 import java.beans.BeanInfo;
 import java.beans.IntrospectionException;