+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - shardkey parameter
 - manager/document/DocumentManager#isForm(InputStream inputStream)
 - the list of image formats available for insertion into the editor has been expanded (tiff)
+- support java 8
 
 ## Changed
 - demo server address changed